refactor: Consolidate tool guard functions into tools.ts, removing toolGuards.ts to enhance code organization and maintainability
